Add unit tests for getBarFromTime and getTimeFromBar
