+++ src/ts/config.ts
@@ -4,12 +4,14 @@ export default {
   "scores": ["modern", "early"],
   
   // Andrew Leslie Cooper
+  
   "audio_prefix": "/audio/ALC/",
   "intro_beats": 2,
   "barno":   [0, 1,      65,    75,  78,    86,    94, 107,   120,   122, 137,   138, 139],
   "bartime": [0, 2.5, 234.5, 274.2, 284, 312.8, 342.4, 388, 437.6, 446.1, 501, 506.4, 512],
 
   // Choir of the Earth
+
   // "audio_prefix": "/audio/CotE/",
   // "intro_beats": 4,
   // "barno":   [0,   1,  40,  65,    81, 102,   109,   120, 128,   133,   137, 138,   139],
